+++ src/Triangle.java
@@ -3,7 +3,7 @@
  */
 public class Triangle {
 
-    public Triangle(int a, int b, int c) {
+    public Triangle(float a, float b, float c) {
 
         if ((a > 0 && b > 0 && c > 0) && (a + b > c && a + c > b && b + c > a)) {
             if (a == b && b == c) {
